# reduce CGO calls when scanning rows